Added hostinfo.py
This script can be used to check an IP address or domain for information. It creates a CSV record for each IP resolution (A and AAAA records) for the domain and gives IP ownership and coarse geo location for the IP address. It uses the local DNS resolver for DNS and it uses a free remote service for IP geo info (rate limited to 150 queries per minute).